refactor: 안쓰는 import제거 및 컨벤션 적용
