+++ src/main/java/com/kakao/cafe/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.kakao.cafe.controller;
 
-import java.net.http.HttpRequest;
 import java.util.List;
 
 import org.slf4j.Logger;
