+++ src/test/java/com/kakao/cafe/repository/article/ArticleRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.kakao.cafe.repository;
+package com.kakao.cafe.repository.article;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import com.kakao.cafe.domain.article.Article;
-import com.kakao.cafe.repository.article.MemoryArticleRepository;
 
 class ArticleRepositoryTest {
