+++ src/test/java/com/kakao/cafe/repository/user/UserRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.kakao.cafe.repository;
+package com.kakao.cafe.repository.user;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -10,8 +10,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import com.kakao.cafe.domain.user.User;
-import com.kakao.cafe.repository.user.MemoryUserRepository;
-import com.kakao.cafe.repository.user.UserRepository;
 
 class UserRepositoryTest {
